Move toggle on/off insights to DisplayOptions
